refactor: 불필요한 코드 삭제, 전반적인 네이밍 변경 (#84)
