+++ api-module/src/main/java/org/devridge/api/common/config/WebSocketConfig.java
@@ -2,6 +2,7 @@ package org.devridge.api.common.config;
 
 import lombok.RequiredArgsConstructor;
 
+import org.devridge.api.common.handler.StompHandler;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
