+++ api-module/src/main/java/org/devridge/api/common/entity/AbstractTimeEntity.java
@@ -1,6 +1,7 @@
-package org.devridge.api.domain.member;
+package org.devridge.api.common.entity;
 
 import lombok.Getter;
+
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
