+++ api-module/src/main/java/org/devridge/api/common/entity/BaseTimeEntity.java
@@ -3,7 +3,6 @@ package org.devridge.api.common.entity;
 import lombok.Getter;
 
 import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.DynamicInsert;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
