+++ api-module/src/main/java/org/devridge/api/common/security/auth/CustomAuthenticationEntryPoint.java
@@ -6,6 +6,7 @@ import org.springframework.security.web.AuthenticationEntryPoint;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
 import java.io.IOException;
 
 public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
