+++ api-module/src/main/java/org/devridge/api/common/security/auth/CustomMemberDetailsService.java
@@ -1,7 +1,9 @@
 package org.devridge.api.common.security.auth;
 
 import lombok.RequiredArgsConstructor;
+
 import org.devridge.api.domain.member.repository.MemberRepository;
+
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
