+++ api-module/src/main/java/org/devridge/api/domain/coffeechat/dto/response/GetAllChatMessage.java
@@ -3,13 +3,13 @@ package org.devridge.api.domain.coffeechat.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import org.devridge.api.common.dto.FindWriterInformation;
+import org.devridge.api.common.dto.UserInformation;
 
 @Getter
 @AllArgsConstructor
 public class GetAllChatMessage {
 
     private Long id;
-    private FindWriterInformation member;
+    private UserInformation member;
     private String content;
 }
