+++ api-module/src/main/java/org/devridge/api/domain/qna/dto/response/GetQnADetailResponse.java
@@ -2,7 +2,7 @@ package org.devridge.api.domain.qna.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
-import org.devridge.api.common.dto.FindWriterInformation;
+import org.devridge.api.common.dto.UserInformation;
 
 import java.time.LocalDateTime;
 
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @Builder
 public class GetQnADetailResponse {
 
-    private FindWriterInformation member;
+    private UserInformation member;
     private String title;
     private String content;
     private Integer views;
